+++ src/promptflow/tests/test_configs/flows/deployment_name_not_enabled/print_env.py
@@ -1,0 +1,17 @@
+from typing import Union
+ 
+ 
+from promptflow import tool
+from promptflow.connections import AzureOpenAIConnection, OpenAIConnection
+
+
+@tool()
+def generate_question(
+    connection: Union[OpenAIConnection, AzureOpenAIConnection],
+    generate_question_prompt: str,
+    deployment_name: str = "",
+    model: str = "",
+    context: str = None,
+    temperature: float = 0.2
+):
+    return {"deployment_name": deployment_name, "model": model}